Revert "MkMenuのitemを切り出して共通化"
This reverts commit ce3679798c5b642dd0e47056f70657e46b382b5e.
